feat(plugins/auth): refactor user auth policies and write auth policies for blacklist function
